Handle corruption in DataStore Preferences gracefully (#5633)
Handle corruption in DataStore Preferences more gracefully.

Tested manually by piping junk data into the datastore files.

When the Settings data store preferences gets corrupt, it will be
replaced by an empty preferences with all null values. This will cause
settings to fall through to sdk defaults until the next settings fetch.
If the Sessions data store preferences gets corrupt, it will lose the
session id until the next session is generated on the next foreground. I
could not force a session to upload will a null session id, but I expect
it to be rare but possible with a multi-process app.